+++ cmd/orc/main.go
@@ -1,10 +1,16 @@
 package main
 
-import "os"
+import (
+	"context"
+	"os"
+	"os/signal"
+)
 
 func main() {
 	prepareCtlOutput(os.Args[1:])
-	if err := rootCmd.Execute(); err != nil {
+	ctx, stop := signal.NotifyContext(context.Background(), os.Interrupt)
+	defer stop()
+	if err := rootCmd.ExecuteContext(ctx); err != nil {
 		if isCtlInvocation(os.Args[1:]) {
 			writeCtlError(os.Stderr, err)
 		}